There are a lot of changes that have been made. Instead of refreshing all of the quality tags each time a part of the tree is clicked, it is generated from a static list. The list can be checked for quality if the run button is hit again.
There is an All Tags area selection. Tags can be edited within the grid and it will be changed within the SQL DB. However there is no notification that shows a change has been made and workbench needs to be refreshed.

If a tag is a large tag that gets broken down into smaller ones, then if one of those smaller ones is bad quality then the larger one is bad quality.

Last thing to update is that == signs have been filtered out including whatever comes following that.